20211113の呉さんの状態に合わせる

+++ app/Http/Controllers/API/V1/MemoController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers\API\V1;
 
 use App\Exceptions\ApiAuthException;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\MemoRequest;
 use App\Models\Memo;
+use App\Protocols\MemoRepositoryProtocol;
 use Illuminate\Contracts\Encryption\DecryptException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Crypt;
@@ -12,6 +14,20 @@ use Illuminate\Support\Str;
 
 class MemoController extends Controller
 {
+    /**
+     * @var MemoRepositoryProtocol
+     */
+    protected $memoRepository;
+
+    /**
+     * MemoController constructor.
+     * @param MemoRepositoryProtocol $memoRepository
+     */
+    public function __construct(MemoRepositoryProtocol $memoRepository)
+    {
+        $this->memoRepository = $memoRepository;
+    }
+
     /**
      * メモ参照API
      * @param Request $request
@@ -42,28 +58,11 @@ class MemoController extends Controller
      * @return \Illuminate\Http\JsonResponse
      * @throws \Illuminate\Validation\ValidationException
      */
-    public function store(Request $request)
+    public function store(MemoRequest $request)
     {
-        $this->validate($request, [
-            'folder_id' => 'nullable|integer',
-            'title' => 'required|string',
-            'contents' => 'required',
-            'is_public' => 'nullable|boolean',
-        ]);
+        $memo = $this->memoRepository->create($request);
 
-        $uuid = Str::uuid()->toString();
-
-        $memo = new Memo();
-        $memo->id = $uuid;
-        $memo->user_id = $request->user() ? $request->user()->id : null;
-        $memo->folder_id = $request->get('folder_id', null);
-        $memo->title = $request->get('title');
-        $memo->contents = $request->get('contents');
-        $memo->is_public = $request->get('is_public', false);
-
-        $memo->save();
-
-        $encryptUUID = Crypt::encryptString($uuid);
+        dd($memo);
 
         return response()->json([
             'id' => $uuid,
@@ -132,7 +131,7 @@ class MemoController extends Controller
     public function delete(Request $request, string $id)
     {
         $key = $request->get('key', null);
-        
+
         if (!$key and !$request->user()) {
             throw new ApiAuthException('no auth');
         }
@@ -147,9 +146,9 @@ class MemoController extends Controller
 
             $memo = Memo::findOrFail($id);
 
-            // if ($memo->user_id !== $request->user()->id) {
-            //     throw new ApiAuthException('no auth', 403);
-            // }
+            if ($memo->user_id !== $request->user()->id) {
+                throw new ApiAuthException('no auth', 403);
+            }
 
             $memo->delete();
